refactor: split signature detection helpers
